remove unconsistent test file
