+++ test/models/vrp_test.rb
@@ -261,30 +261,5 @@ module Models
       assert vrp.routes.first.day_index
       assert_equal 10, vrp.routes.first.day_index
     end
-
-    def test_reject_if_several_visits_but_no_schedule_provided
-      # If this test fail then it probably returns 'Wrong number of visits returned in result'
-      # This makes sense, since we do not expand the problem if no schedule is provided,
-      # therefore there is a gap between expected and returned number of visits
-      vrp = VRP.toy
-      vrp[:services][0][:visits_number] = 10
-
-      assert_raises OptimizerWrapper::DiscordantProblemError do
-        TestHelper.create(vrp)
-      end
-
-      vrp[:configuration][:schedule] = { range_indices: { start: 0, end: 10 } }
-      assert TestHelper.create(vrp) # no raise when schedule is provided
-
-      vrp = VRP.pud
-      vrp[:shipments][0][:visits_number] = 10
-
-      assert_raises OptimizerWrapper::DiscordantProblemError do
-        TestHelper.create(vrp)
-      end
-
-      vrp[:configuration][:schedule] = { range_indices: { start: 0, end: 10 } }
-      assert TestHelper.create(vrp) # no raise when no schedule is provided
-    end
   end
 end
